refactor(rome_js_anaylze): introduce StaticValue enum for checking value of the JSX attributes or JS expressions (#4342)

+++ crates/rome_js_analyze/src/semantic_analyzers/style/no_shouty_constants.rs
@@ -66,7 +66,7 @@ fn is_id_and_string_literal_inner_text_equal(
         .as_js_string_literal_expression()?;
     let literal_text = literal.inner_string_text().ok()?;
 
-    if id_text.len() != usize::from(literal_text.len()) {
+    if id_text.len() != literal_text.len() {
         return None;
     }
